pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sc-system-ai"
version = "0.0.2"
description = "A sample Python project"
readme = "README.md"
dependencies = [
    "python-dotenv",
    "langchain",
    "langchain-community",
    "langchain-openai",
    "langsmith",
    "langgraph",
    "duckduckgo-search",
    "azure-cosmos",
    "pypdf",
]

[tool.setuptools.packages.find]
where = ["src"]